Adds force update option for WDPA downloads
Adds a `force_update` parameter to `make_protected_areas` to allow users to force a re-download of the WDPA database, even if a local version exists and is up-to-date. This ensures the latest data is used when needed.

Also, updates the log message to clarify when the WDPA will be updated. It now specifies that updates happen only when `force_update = TRUE`.

Updates the included example code to use `force_update = TRUE` for demonstration.